% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate_helpers.R
\name{elapsed}
\alias{elapsed}
\title{Compute elapsed time between two dates}
\usage{
elapsed(start_datetime, end_datetime, units = "years")
}
\arguments{
\item{start_datetime}{The beginning of the interval}

\item{end_datetime}{The end of the interval}

\item{units}{The units for the interval, e.g. "seconds" or "years"}
}
\value{
a floating-point quantity that is the number of \code{units} between \code{start_datetime} and \code{end_datetime}.
}
\description{
Compute elapsed time between two dates
}
\examples{
elapsed("2020-02-29", "2025-04-30", "years")
}
