% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HEx_COLORS}
\alias{HEx_COLORS}
\title{Branding colors from Higher Expectations for Racine County}
\format{
a list of color names and hex codes
}
\usage{
HEx_COLORS
}
\description{
Branding colors from Higher Expectations for Racine County
}
\keyword{datasets}
