% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NSC_CLASS_YEARS}
\alias{NSC_CLASS_YEARS}
\title{Lookup table of one-letter codes for postsecondary class years}
\format{
A two-column table of abbreviations and glosses
\describe{
\item{Value}{one-letter codes for different class years, e.g. "J."}
\item{Meaning}{a short description of what the code means, e.g. "Junior."}
}
}
\usage{
NSC_CLASS_YEARS
}
\description{
Lookup table of one-letter codes for postsecondary class years
}
\keyword{datasets}
