% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NSC_ENROLLMENT_STATUSES}
\alias{NSC_ENROLLMENT_STATUSES}
\title{Lookup table of one-letter codes for students' enrollment statuses}
\format{
A two-column table of abbreviations and glosses
\describe{
\item{Value}{one-letter codes for different statuses, e.g. "G"}
\item{Meaning}{a short description of what the code means, e.g. "Graduated."}
}
}
\usage{
NSC_ENROLLMENT_STATUSES
}
\description{
Lookup table of one-letter codes for students' enrollment statuses
}
\keyword{datasets}
