% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NSC_GRADUATION_STATUS_LEVELS}
\alias{NSC_GRADUATION_STATUS_LEVELS}
\title{Either graduated, in, or passed the graduation window}
\format{
An object of class \code{character} of length 3.
}
\usage{
NSC_GRADUATION_STATUS_LEVELS
}
\description{
Either graduated, in, or passed the graduation window
}
\keyword{datasets}
