% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NSC_STUDENT_TRACKER_FIELDS}
\alias{NSC_STUDENT_TRACKER_FIELDS}
\title{Specification for the layout of Student Tracker reports}
\format{
A five-column table
\describe{
\item{COLUMN POSITION}{letter codes for column positions, e.g. "A" for column 1.}
\item{FIELD NAME}{A short, human-readable label for each column in the report.}
\item{MAX. LENGTH}{The number of characters in the field}
\item{TYPE}{A brief alphanumeric code describing the data type}
\item{DESCRIPTION}{A long text description of the field}
}
}
\source{
https://help.studentclearinghouse.org/sths/wp-content/uploads/StudentTracker-3.0-Student-level-Detail-Report-Mapping.pdf
}
\usage{
NSC_STUDENT_TRACKER_FIELDS
}
\description{
Specification for the layout of Student Tracker reports
}
\keyword{datasets}
