% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NSC_STUDENTS_SPEC}
\alias{NSC_STUDENTS_SPEC}
\title{A column specification for importing student data from NSC.}
\format{
A list of columns names with data type abbreviations
}
\usage{
NSC_STUDENTS_SPEC
}
\description{
A column specification for importing student data from NSC.
}
\seealso{
\code{\link[readr:cols]{readr::cols()}}
}
\keyword{datasets}
