% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPEC_FOR_TOP25_INSTITUTIONS_BYHSCLASS}
\alias{SPEC_FOR_TOP25_INSTITUTIONS_BYHSCLASS}
\title{Column types for data about enrollment at the top 25 college destinations}
\format{
A list of one-character column type codes, named by column
}
\usage{
SPEC_FOR_TOP25_INSTITUTIONS_BYHSCLASS
}
\description{
Column types for data about enrollment at the top 25 college destinations
}
\keyword{datasets}
