% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_status.R
\name{student_status}
\alias{student_status}
\title{Convert graduation and activity fields into graduation status}
\usage{
student_status(graduation, activity)
}
\arguments{
\item{graduation}{a text field about whether or not a student has graduated.}

\item{activity}{a text field about the student's enrollment activity}
}
\value{
one of "Graduated," "Within Graduation Window," or "Past Graduation Window."
}
\description{
Convert graduation and activity fields into graduation status
}
