% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WI_SE_OTHER_HIGHER_EDS}
\alias{WI_SE_OTHER_HIGHER_EDS}
\title{Lookup table of categorical values for uncommon postsecondary choices}
\format{
A two-column table of higher ed. names and categories
\describe{
\item{College Name}{Verbose National Student Clearinghouse names for higher eds.}
\item{Institution}{a short factor level that the higher ed. belongs to.}
}
#' ,
}
\usage{
WI_SE_OTHER_HIGHER_EDS
}
\description{
Lookup table of categorical values for uncommon postsecondary choices
}
\keyword{datasets}
